# swagger-api.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cocktail Ordering System API",
    "version": "1.0.0",
    "description": "A complete backend API for a cocktail ordering system with Express.js, Node.js, and MongoDB. Features JWT authentication, Paystack payment integration, and comprehensive admin management.",
    "contact": {
      "name": "Clyp AI",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    },
...
  },
...
  "paths": {
...
    "/inventory": {
      "get": {
        "summary": "Get all inventory items with filtering (Admin only)",
        "tags": [
          "Inventory Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "sufficient",
                "low",
                "critical",
                "out_of_stock"
              ]
            },
            "description": "Filter by stock status"
          },
          {
            "in": "query",
            "name": "lowStock",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter for low stock items only"
          },
          {
            "in": "query",
            "name": "criticalStock",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter for critical stock items only"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Number of items to return"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "Inventory items retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "inventory": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Inventory"
                      }
                    },
                    "summary": {
                      "type": "object",
                      "properties": {
                        "totalItems": {
                          "type": "integer"
                        },
                        "lowStock": {
                          "type": "integer"
                        },
                        "criticalStock": {
                          "type": "integer"
                        },
                        "outOfStock": {
                          "type": "integer"
                        }
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        },
                        "pages": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Admin access required"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create inventory item for a cocktail (Admin only)",
        "tags": [
          "Inventory Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "cocktail",
                  "currentStock",
                  "minimumStock",
                  "maximumStock"
                ],
                "properties": {
                  "cocktail": {
                    "type": "string",
                    "description": "Cocktail ID"
                  },
                  "currentStock": {
                    "type": "number",
                    "minimum": 0
                  },
                  "minimumStock": {
                    "type": "number",
                    "minimum": 0
                  },
                  "maximumStock": {
                    "type": "number",
                    "minimum": 1
                  },
                  "unit": {
                    "type": "string",
                    "enum": [
                      "bottles",
                      "liters",
                      "gallons",
                      "pieces",
                      "servings"
                    ]
                  },
                  "costPerUnit": {
                    "type": "number",
                    "minimum": 0
                  },
                  "supplier": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "contact": {
                        "type": "object",
                        "properties": {
                          "phone": {
                            "type": "string"
                          },
                          "email": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "alertSettings": {
                    "type": "object",
                    "properties": {
                      "isEnabled": {
                        "type": "boolean"
                      },
                      "frequency": {
                        "type": "string",
                        "enum": [
                          "immediate",
                          "daily",
                          "weekly",
                          "monthly"
                        ]
                      },
                      "alertThreshold": {
                        "type": "number",
                        "minimum": 0
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Inventory item created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "inventory": {
                      "$ref": "#/components/schemas/Inventory"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or cocktail already has inventory"
          },
          "404": {
            "description": "Cocktail not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/inventory/{id}/restock": {
      "post": {
        "summary": "Restock inventory item (Admin only)",
        "tags": [
          "Inventory Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Inventory item ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "quantity"
                ],
                "properties": {
                  "quantity": {
                    "type": "number",
                    "minimum": 1,
                    "description": "Quantity to add to stock"
                  },
                  "cost": {
                    "type": "number",
                    "minimum": 0,
                    "description": "Total cost of restock"
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 500,
                    "description": "Restock notes"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inventory restocked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "inventory": {
                      "$ref": "#/components/schemas/Inventory"
                    },
                    "restockInfo": {
                      "type": "object",
                      "properties": {
                        "quantityAdded": {
                          "type": "number"
                        },
                        "previousStock": {
                          "type": "number"
                        },
                        "newStock": {
                          "type": "number"
                        },
                        "totalCost": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Inventory item not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/inventory/report": {
      "get": {
        "summary": "Generate inventory report (Admin only)",
        "tags": [
          "Inventory Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Inventory report generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "report": {
                      "type": "object",
                      "properties": {
                        "timestamp": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "summary": {
                          "type": "object"
                        },
                        "lowStock": {
                          "type": "array"
                        },
                        "criticalStock": {
                          "type": "array"
                        },
                        "outOfStock": {
                          "type": "array"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}